chore: move test api cases from expect/ to test-api/
