+++ tests/test-api/fixtures/timeout.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from '@rstest/core';
-import { sleep } from '../../../scripts';
+import { sleep } from '../../scripts';
 
 describe('level A', () => {
   it('it in level A', async () => {
